# Faster saving (delay progress UI for 1s)