--- a/editor/progress_dialog.h
+++ b/editor/progress_dialog.h
@@ -76,6 +76,7 @@ class ProgressDialog : public Popup {
 		VBoxContainer *vb;
 		ProgressBar *progress;
 		Label *state;
+		uint64_t showtime;
 	};
 	HBoxContainer *cancel_hb;
 	Button *cancel;