@reduz 2017-12-21
Not going to remove the progress bar for just 100ms...

On Thu, Dec 21, 2017 at 3:34 AM, Bernhard Liebl <notifications@github.com>
wrote:

> EditorNode::_save_scene_with_preview currently takes about 500 ms on my
> iMac (optimized build).
>
> About 40% of that time is spent in ui code (displaying the progress bar
> and handling other ui events in Main::iteration()). If I disable the
> progress bar display during save, the time goes down to about 320 ms.
>
> So I suggest hiding the progress bar only if the save takes less than 1s,
> since:
>
> (a) on many systems, the save progress bar is merely a flicker (like 100
> ms) and a bit annoying IMHO
> (b) showing the progress bar substantially increases the save scene time,
> and as such the time to start the runner during game development
> (c) the save progress bar cannot be cancelled and so there is nothing lost
> if the ui gets unresponsive during save for up to 1s
>
> Here are some concrete measurements of EditorNode::_save_scene_with_
> preview in milliseconds. Before the change:
>
> _save_scene_with_preview 549
> _save_scene_with_preview 514
> _save_scene_with_preview 485
> _save_scene_with_preview 481
> _save_scene_with_preview 538
> _save_scene_with_preview 474
> _save_scene_with_preview 470
> _save_scene_with_preview 456
> _save_scene_with_preview 538
>
> After the change:
>
> _save_scene_with_preview 353
> _save_scene_with_preview 324
> _save_scene_with_preview 308
> _save_scene_with_preview 345
> _save_scene_with_preview 321
> _save_scene_with_preview 306
> _save_scene_with_preview 332
>
> This PR checks on each progress step if 1s has passed; if not, the
> progress bar is not yet shown.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/godotengine/godot/pull/14891
> Commit Summary
>
>    - Faster saving (delay progress UI for 1s)
>
> File Changes
>
>    - *M* editor/editor_node.cpp
>    <https://github.com/godotengine/godot/pull/14891/files#diff-0> (6)
>    - *M* editor/editor_node.h
>    <https://github.com/godotengine/godot/pull/14891/files#diff-1> (6)
>    - *M* editor/progress_dialog.cpp
>    <https://github.com/godotengine/godot/pull/14891/files#diff-2> (7)
>    - *M* editor/progress_dialog.h
>    <https://github.com/godotengine/godot/pull/14891/files#diff-3> (3)
>
> Patch Links:
>
>    - https://github.com/godotengine/godot/pull/14891.patch
>    - https://github.com/godotengine/godot/pull/14891.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/godotengine/godot/pull/14891>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AF-Z2xWxogrWsHeG7d7FyquuYOgKhTK9ks5tCfvYgaJpZM4RJaGx>
> .
>