@poke1024 2017-12-21
Changed the PR's title as its original wording only talks about performance and is thus missing the main point.

Should a progress indicator be shown for a save operation needing < 1 s? I think no. The common UI/UX lore is to show progress bars for operations >= 1 s:

- https://msdn.microsoft.com/en-us/library/windows/desktop/dn742481(v=vs.85).aspx: "Use an activity indicator when an action can't be performed within a second"
- Jakob Nielsen: https://www.nngroup.com/articles/response-times-3-important-limits/: "Normally, no special feedback is necessary during delays of more than 0.1 but less than 1.0 second"