@hpvb 2017-12-22
This makes sense to me, with the new explanation.