@letheed 2017-12-22
Personally, I like having the visual feedback when I save.