@mhilbrunner 2017-12-22
Having a progress bar for sub-second actions seems not useful, IMO.