@codingHahn 2017-12-24
@reduz What do you think about it now with the new explanation?