@reduz 2017-12-24
You can't tell in advance how long a scene will take to save, and for a few milliseconds this is not worth it.  If many of you for some bizarre option don't like this dialog, we could make it optionally disabled but it won't go away.