@star-tek-mb 2017-12-27
No need for this hack. Keep code simplier.