@poke1024 2017-12-28
Changed this into an editor setting currently called "Show Progress Bar For Longer Tasks Only". It's disabled by default. If you enable it, the progress bar will only show after 1s; we might have to fine tune this for some operations that call the progress update/main iteration very seldom.