@akien-mga 2018-01-04
For the reference, I don't think this should be an editor setting. The answer to a lack of consensus should not always be "let's make it an option".