@reduz 2018-05-08
Closing, because I don't think this will work as intended in practice.